requirements.txt

numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0
json5>=0.9.0
tqdm>=4.62.0
openai>=1.0.0
nltk>=3.7
jieba>=0.42.1
transformers>=4.20.0
tokenizers>=0.13.0
torch>=1.12.0
librosa>=0.9.0
soundfile>=0.10.0
pyloudnorm>=0.1.0
onnxruntime>=1.12.0
demucs>=3.0.0
pyannote.audio>=2.1.0
openai-whisper>=20230314
jiwer>=2.3.0
pydub>=0.25.0
huggingface_hub>=0.10.0
yt-dlp>=2023.1.6 